+++ CHANGELOG.md
@@ -614,6 +614,7 @@
 * Added /zone_tick and /zone_tick_remove to set the chat tick rate of a zone, so all players in an area part of the zone see messages with the same chat tick rate, or their own chat tick rate
 * Made /switch indicate the target character, regardless of whether the switch was successful or not
 * Made /zone_watch return a more specific error if the player is already watching the target zone
+* Added /ignore and /unignore to toggle ignoring IC messages from players
 * Fixed scream_range in area list yaml files not supporting the keyword <ALL> to indicate all areas should be able to receive a scream coming from a particular area
 * Fixed scream_range in area list yaml files not checking if the areas a scream can reach to from a particular area exist
 * Fixed /scream, /whisper and /party_whisper raising errors if a message was sent to a deafened player with a bypass message starter. They now sent messages but filtered